refactor(vnext): finish gateway/ — channel-history (#623 step 5)
Pure relocate, no logic change. Completes the L2 gateway/ layer's
inbound side.

- src/core/channel-history.{ts,test.ts} → src/vnext/gateway/

Three modules originally in this PR were pulled back to src/core/ for
later steps:
- tool-result-truncation: agent context-window concern, not gateway
- assistant-output-directives: pure system-prompt fragment, belongs with
  workspace.ts in agent/
- silent-reply: behavior questionable; reassess before relocating

External importer updated: plugins/discord/discord.ts.

Final gateway/ shape (6 modules):
- routing decisions: bindings, mention, session-keys
- inbound primitives: dedupe, debounce
- channel state: channel-history
